pyproject.toml
[project]
name = "project-root-0-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "pandas>=2.3.3",
    "pypdf>=6.6.0",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.53.0",
]
